listagem parcialmente pronta

+++ src/controllers/ClientsController.java
@@ -1,13 +1,14 @@
 package controllers;
 
 import com.jfoenix.controls.JFXTextField;
+import javafx.concurrent.Task;
 import javafx.event.ActionEvent;
+import javafx.event.EventHandler;
 import javafx.fxml.Initializable;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
 import logic.Client;
-import logic.Product;
 import services.ClientService;
 
 import java.net.URL;
@@ -32,21 +33,27 @@ public class ClientsController extends MainController implements Initializable {
         adressColumn.setCellValueFactory(new PropertyValueFactory<>("Address"));
         emailColumn.setCellValueFactory(new PropertyValueFactory<>("Email"));
         phoneColumn.setCellValueFactory(new PropertyValueFactory<>("Phone"));
+        mainTableView.getItems().add(null);
 
-        try {
-            clientService = new ClientService();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
+        showLoader();
+        Task retriveClientsTask = new Task() {
 
-        try {
-            this.clientsList = clientService.getAllClients();
-        } catch (Exception e) {
-            showMsg(e.getMessage());
-            e.printStackTrace();
-        }
+            @Override
+            protected List<Client> call() throws Exception {
+                clientService = new ClientService();
+                return clientService.getAllClients();
+            }
 
-        showOnTable();
+            @Override
+            protected void succeeded() {
+                clientsList = (List<Client>) getValue();
+                hideLoader();
+                showOnTable();
+            }
+        };
+        Thread t = new Thread(retriveClientsTask);
+        t.setDaemon(true);
+        t.start();
 
         //Definir que quando o enter for pressionado o filtro ocorra
         searchInput.setOnAction(this::filterSearch);
@@ -56,6 +63,7 @@ public class ClientsController extends MainController implements Initializable {
 
 
     private void showOnTable() {
+        mainTableView.getItems().clear();
         this.clientsList.stream()
                 .sorted(Client::compareTo)
                 .forEach(client -> mainTableView.getItems().add(client));
@@ -79,23 +87,35 @@ public class ClientsController extends MainController implements Initializable {
     }
 
     public void deleteClients(ActionEvent actionEvent) {
-        try {
-            Client c = mainTableView.getSelectionModel().getSelectedItem();
-            mainTableView.getItems().removeAll(c);
-            clientService.deleteClient(c);
-            clientsList.remove(c);
-        } catch (Exception e) {
-            e.printStackTrace();
+        showLoader();
+        if (mainTableView.getSelectionModel().getSelectedItem() != null) {
+            try {
+                Client c = mainTableView.getSelectionModel().getSelectedItem();
+                mainTableView.getItems().removeAll(c);
+                clientService.deleteClient(c);
+                clientsList.remove(c);
+                hideLoader();
+            } catch (Exception e) {
+                showMsg("Ocorreu um erro" + e.getMessage());
+                e.printStackTrace();
+                hideLoader();
+            }
+        } else {
+            showMsg("Selecione alguma linha para continuar");
+            hideLoader();
         }
     }
 
     public void editClients(ActionEvent actionEvent) {
-        try {
-            actualClient = mainTableView.getSelectionModel().getSelectedItem();
-            clearMainArea();
-            loadCenterUI("/fxml/RegisterClients.fxml");
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
+        if (mainTableView.getSelectionModel().getSelectedItem() != null) {
+            try {
+                actualClient = mainTableView.getSelectionModel().getSelectedItem();
+                clearMainArea();
+                loadCenterUI("/fxml/RegisterClients.fxml");
+            } catch (Exception e) {
+                showMsg("Ocorreu um erro" + e.getMessage());
+                e.printStackTrace();
+            }
+        } else showMsg("Selecione alguma linha para continuar");
     }
 }
